pipeline/runtime-requirements.txt
# Generated by KFP.
boto3
botocore
pillow
torch
torchvision